+++ Sources/SwiftKit/Authentication/CachedAuthenticationService.swift
@@ -10,26 +10,35 @@ import Foundation
 
 /**
  This protocol can be implemented by services that can cache
- authentication results, to avoid having to perform the real
- authentication operation every time.
+ an authentication result, to avoid having to perform a real
+ authentication operation if a successful authentication has
+ already been performed.
  
  For instance, you can reduce the number of times users have
  to perform biometric authentication to access critical data.
- Note that you can never rely on the cache implementation to
- clear this data. Always call `resetAuthentication(for:)` as
- soon as the state can be considered to become invalid, e.g.
- when the app resigns its active state.
+ 
+ Note that you can't rely on a cached authentication service
+ to clear its cached state. Call `resetUserAuthentication()`
+ or `resetUserAuthentication(for:)` as soon as this state is
+ considered to be invalid, e.g. when your app is send to the
+ background and a new user can open the app at a later time.
  */
 public protocol CachedAuthenticationService: AuthenticationService {
     
     /**
-     Check if a certain authentication is authenticated, e.g. if
-     the cached authentication state is still valid.
+     Check if the user has already been authenticated for an
+     authentication type.
      */
-    func isAlreadyAuthenticated(for auth: Authentication) -> Bool
+    func isUserAuthenticated(for auth: Authentication) -> Bool
     
     /**
-     Reset a cached authentication state for a specific auth.
+     Reset the user's entire authentication state.
      */
-    func resetAuthenticationCache(for auth: Authentication)
+    func resetUserAuthentication()
+    
+    /**
+     Reset the user's authentication state for a single type
+     of authentication.
+     */
+    func resetUserAuthentication(for auth: Authentication)
 }