+++ Release Notes.md
@@ -1,9 +1,14 @@
 # Release notes
 
 
+## 0.0.2
+
+This release adjusts [Authentication][Authentication] service signatures.
+
+
 ## 0.0.1
 
-This very first release adds [Authentication][Authentication] functionality to `SwiftKit`.
+This release adds [Authentication][Authentication] functionality to `SwiftKit`.
 
 
 [Authentication]: Readmes/Authentication.md